move fuse comm to tensor fusion helper
