+++ python/paddle/distributed/fleet/meta_parallel/pipeline_parallel.py
@@ -37,10 +37,10 @@ else:
     from .pp_utils import p2p_communication as p2p
 
 from paddle.distributed.fleet.utils.tensor_fusion_helper import (
+    HOOK_ACTION,
+    FusedCommBuffer,
     assign_group_by_size,
 )
-
-from .pp_utils.utils import HOOK_ACTION, FusedCommBuffer
 
 __all__ = []
