+++ python/paddle/distributed/fleet/utils/tensor_fusion_helper.py
@@ -17,7 +17,14 @@ from collections import OrderedDict
 import numpy as np
 
 import paddle
+from paddle.framework import base as imperative_base
 from paddle.framework import core
+
+
+class HOOK_ACTION:
+    ALL_REDUCE = 0
+    REDUCE = 1
+
 
 alignment = {
     "gpu": 256,
@@ -201,3 +208,118 @@ def fused_parameters(parameters, use_main_grad):
         all_fused += other
 
     return decay_fused, all_fused
+
+
+class FusedCommBuffer:
+    def __init__(self, id, params, comm_group, acc_steps=1, act=None, dst=-1):
+        self._id = id
+        self._params = params
+        self._acc_steps = acc_steps
+        self._comm_group = comm_group
+
+        self.use_main_grad = hasattr(self._params[0], "main_grad")
+
+        self._task = None
+        self._params_step_dict = {}
+        self._params_checked_in = 0
+        self._params_to_addr = {}
+
+        self._act = act
+        if self._act == HOOK_ACTION.ALL_REDUCE:
+            assert dst == -1
+        elif self._act == HOOK_ACTION.REDUCE:
+            assert dst != -1
+        else:
+            raise ValueError(
+                "The act should be allreudce for dp or reduce for sharding."
+            )
+        self._dst = dst
+
+        self._init_step_dict()
+
+        self.grad_storage = flatten_dense_tensors(
+            self._params,
+            use_main_grad=self.use_main_grad,
+            fuse_param=False,
+            warp_buffer=False,
+        ).buffer
+
+        self._record_addr()
+
+    def _record_addr(self):
+        for param in self._params:
+            addr = (
+                param.main_grad.data_ptr()
+                if self.use_main_grad
+                else param.grad.data_ptr()
+            )
+            self._params_to_addr[param.name] = addr
+
+    def _init_step_dict(self):
+        for p in self._params:
+            self._params_step_dict[p.name] = 0
+
+    def _reset_params_checked_in(self):
+        self._task = None
+        self._init_step_dict()
+        self._params_checked_in = 0
+
+    @property
+    def _all_params_checked_in(self):
+        return (
+            len(self._params) == self._params_checked_in
+            and len(self._params_step_dict) == 0
+        )
+
+    def add_grad(self, param):
+        assert param.name in self._params_step_dict
+        current_ptr = (
+            param.main_grad.data_ptr()
+            if self.use_main_grad
+            else param.grad.data_ptr()
+        )
+        if self._params_to_addr[param.name] != current_ptr:
+            raise ValueError(
+                "The address of the grad/main_grad of the param has been changed during training, "
+                "which is not allowed for dp/sharding overlap with pp. "
+                "This may be caused by some non-inplace operations on the grad/main_grad. "
+                "Please use the inplace version of the operations or disable the overlapping."
+            )
+
+        self._params_step_dict[param.name] += 1
+
+        if self._params_step_dict[param.name] == self._acc_steps:
+            self._params_checked_in += 1
+            self._params_step_dict.pop(param.name)
+
+        if self._all_params_checked_in:
+            self._comm_grads()
+
+    @imperative_base.no_grad
+    def _comm_grads(self):
+        assert self._all_params_checked_in
+
+        if self._act == HOOK_ACTION.ALL_REDUCE:
+            task = paddle.distributed.all_reduce(
+                self.grad_storage, group=self._comm_group, sync_op=False
+            )
+
+        elif self._act == HOOK_ACTION.REDUCE:
+            task = paddle.distributed.reduce(
+                self.grad_storage,
+                dst=self._dst,
+                group=self._comm_group,
+                sync_op=False,
+            )
+
+        self._task = task
+
+    @imperative_base.no_grad
+    def scale_and_split_grads(self):
+        assert self._task is not None
+        self._task.wait()
+
+        scale_factor = 1.0 / self._comm_group.nranks
+        self.grad_storage.scale_(scale_factor)
+
+        self._reset_params_checked_in()
